tamed "bpy.context.selected_objects" list
With 2 objects the way Blender sorts the list obtained by "bpy.context.selected_objects" is not helping when there is a need to predict to which of the selected objects the vertex is added. This way the new vertex is always added to the mesh on which the edge was selected first.